+++ server/repository/db-updates.go
@@ -11,7 +11,7 @@ import (
 )
 
 func RunDBSchemaUpdates() {
-	targetVersion := 42
+	targetVersion := 43
 	curVersion, err := GetSettingsRepository().GetGlobalInt(SettingDatabaseVersion.Name)
 	log.Printf("Initializing database with schema version %d (current: %d) …\n", targetVersion, curVersion)
 	if err != nil {
@@ -32,7 +32,6 @@ func RunDBSchemaUpdates() {
 		GetSettingsRepository(),
 		GetRecurringBookingRepository(),
 		GetRefreshTokenRepository(),
-		GetDebugTimeIssuesRepository(),
 		GetSpaceAttributeRepository(),
 		GetSpaceAttributeValueRepository(),
 		GetMailLogRepository(),
@@ -45,6 +44,13 @@ func RunDBSchemaUpdates() {
 	for _, repository := range repositories {
 		repository.RunSchemaUpgrade(curVersion, targetVersion)
 	}
+
+	if curVersion < 43 {
+		if _, err := GetDatabase().DB().Exec("DROP TABLE IF EXISTS debug_time_issues"); err != nil {
+			panic(err)
+		}
+	}
+
 	GetSettingsRepository().SetGlobal(SettingDatabaseVersion.Name, strconv.Itoa(targetVersion))
 	SetGlobalInstallID()
 }