+++ server/repository/test/booking-repository_test.go
@@ -158,25 +158,7 @@ func TestBookingRepositoryGetBookingsRequiringApproval(t *testing.T) {
 	}
 	GetGroupRepository().Create(group)
 	GetGroupRepository().AddMembers(group, []string{adminUser.ID})
-	location := &Location{
-		Name:           "Location 1",
-		OrganizationID: org.ID,
-	}
-	if err := GetLocationRepository().Create(location); err != nil {
-		t.Fatalf("Expected nil error, but got %s\n%s", err, debug.Stack())
-	}
-	space := &Space{
-		Name:       "H234",
-		X:          50,
-		Y:          100,
-		Width:      200,
-		Height:     300,
-		Rotation:   90,
-		LocationID: location.ID,
-	}
-	if err := GetSpaceRepository().Create(space); err != nil {
-		t.Fatalf("Expected nil error, but got %s\n%s", err, debug.Stack())
-	}
+	_, space := CreateTestLocationAndSpace(org)
 	if err := GetSpaceRepository().AddApprovers(space, []string{group.ID}); err != nil {
 		t.Fatalf("Expected nil error, but got %s\n%s", err, debug.Stack())
 	}
@@ -192,14 +174,7 @@ func TestBookingRepositoryGetBookingsRequiringApproval(t *testing.T) {
 	}
 	CheckTestInt(t, 0, count)
 
-	booking := &Booking{
-		UserID:   user.ID,
-		SpaceID:  space.ID,
-		Enter:    time.Now().Add(2 * time.Hour),
-		Leave:    time.Now().Add(4 * time.Hour),
-		Approved: false,
-	}
-	GetBookingRepository().Create(booking)
+	booking := CreateTestBooking9To5(user, space, 1)
 
 	bookings, err = GetBookingRepository().GetBookingsRequiringApproval(adminUser.ID)
 	if err != nil {
@@ -215,6 +190,19 @@ func TestBookingRepositoryGetBookingsRequiringApproval(t *testing.T) {
 
 	booking.Approved = true
 	GetBookingRepository().Update(booking)
+
+	bookings, err = GetBookingRepository().GetBookingsRequiringApproval(adminUser.ID)
+	if err != nil {
+		t.Fatalf("Expected nil error, but got %s\n%s", err, debug.Stack())
+	}
+	CheckTestInt(t, 0, len(bookings))
+	count, err = GetBookingRepository().GetBookingsCountRequiringApproval(adminUser.ID)
+	if err != nil {
+		t.Fatalf("Expected nil error, but got %s\n%s", err, debug.Stack())
+	}
+	CheckTestInt(t, 0, count)
+
+	CreateTestBooking9To5(user, space, -2)
 
 	bookings, err = GetBookingRepository().GetBookingsRequiringApproval(adminUser.ID)
 	if err != nil {