+++ ui/src/util/DateUtil.ts
@@ -95,14 +95,6 @@ export default class DateUtil {
     );
   }
 
-  static isInFuture(date: Date): boolean {
-    return this.convertToUTC(date) > new Date();
-  }
-
-  static isAfterToday(date: Date): boolean {
-    return this.isInFuture(date) && !this.isToday(date);
-  }
-
   static isInPast(date: Date): boolean {
     return this.convertToUTC(date) < new Date();
   }
@@ -276,6 +268,10 @@ export default class DateUtil {
 
   static getNowFakeUTC(): Date {
     return this.convertToFakeUTCDate(new Date());
+  }
+
+  static hoursToDay(hours: number) {
+    return Math.floor(hours / 24);
   }
 
   /**